refactor: 권한에 따른 패키지 분류
